Crate of

A pure-Rust #![no_std] crate for parsing Flattened Devicetrees, with the goal of having a very ergonomic and idiomatic API.

Re-exports

pub use kernel_nodes::*;

Modules

kernel_nodes

Structs

Cpu

Represents a /cpus/cpu* node with specific helper methods

MachineFdt

OfPhandleArgs

Functions

bootargs

Bootargs from the FDT

cpus

CPU nodes in the FDT

fdt_available⚠

Whether the FDT is available

fdt_size

Returns the size of the FDT

find_compatible_node

Searches for a node which contains a compatible property and contains one of the strings inside of with

find_phandle

Find a phandle in the FDT

get_fdt_ptr

init_fdt_ptr⚠

Safety

machin_name

Root Node found model or first compatible

memory_nodes

All memory nodes in the FDT

of_device_is_available

Whether the device pointed by the node is available

of_parse_phandle_with_args

Parse a phandle with arguments

of_property_read_u32

Read a u32 property from a node

pcsi

Psci node in the FDT

Type Aliases

OfNode